Refactored and create DisplayPerson class
